# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Book Management API
  version: 1.3.1
  contact: {}
paths:
...
  /api/v1/books/update/{id}:
    patch:
      tags:
        - Bookstore
      summary: Update book
      description: >-
        The API endpoint allows you to update a specific book.
      operationId: updateBook
      parameters:
        - name: x-access-token
          in: header
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: New Book
                author:
                  type: string
                  example: test@123
                publicationYear:
                  type: number
                  example: 2024
      responses:
        "201":
          description: Update Book
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: "500"
            Date:
              schema:
                type: string
                example: "Fri, 30 Jun 2023 06:44:09 GMT"
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-RateLimit-Limit:
              schema:
                type: string
                example: "500"
            X-RateLimit-Remaining:
              schema:
                type: string
                example: "495"
            X-RateLimit-Reset:
              schema:
                type: string
                example: "1714049714"
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Successfully completed the request
                  data:
                    type: string
                    example: "Book data updated successfully"
                  error:
                    type: object
                    example: {}
              examples:
                Update Book:
                  value:
                    status: 200
                    success: true
                    message: Successfully completed the request.
                    data: Book data updated successfully
                    error: {}
...